+++ src/modules/constructor/foreignConstructor.js
